requirements.txt
dj-database-url==0.5.0
Django==2.2
django-filter==2.1.0
django-localflavor==2.1
django-phone-field==1.7.1
python-decouple==3.1
pytz==2019.1
sqlparse==0.3.0
